Setup proper module structure
