+++ pyproject.toml
@@ -35,14 +35,8 @@ optional = true
 pytest = "^7.1.2"
 pytest-cov = "^3.0"
 
-
 [tool.isort]
-multi_line_output=3
-include_trailing_comma=true
-force_grid_wrap=0
-use_parentheses=true
-line_length=88
-balanced_wrapping = 1
+profile = "black"
 default_section = "THIRDPARTY"
 known_first_party = "yamjinx"
 known_third_party= "tests"
