+++ yamjinx/containers/settings.py
@@ -1,0 +1,8 @@
+from attr import frozen
+
+
+@frozen
+class IndentConfig:
+    mapping: int
+    sequence: int
+    offset: int
