+++ yamjinx/loader/__init__.py
@@ -1,8 +1,4 @@
 from yamjinx.loader.preprocessor import translate_config_flags
 from yamjinx.loader.validator import validate_content
 
-
-__all__ = [
-    "translate_config_flags",
-    "validate_content"
-]
+__all__ = ["translate_config_flags", "validate_content"]
