+++ yamjinx/loader/validator.py
@@ -1,10 +1,12 @@
 from typing import Any
+
 from yamjinx.constants import (
-    TOGGLE_KEY_PREFIX,
-    NON_EXISTING_STR,
     DEDUPLICATOR,
+    NON_EXISTING_STR,
+    TOGGLE_KEY_PREFIX,
     TOGGLE_TAG_PREFIX,
 )
+
 
 def validate_content(data: Any) -> None:
     """Checks that internally used fields do not exist in input yaml file"""
